refactor routes module into separate routes and controllers modules
