+++ app/server/src/types/app.ts
@@ -1,0 +1,6 @@
+import {Application} from "express";
+
+export type BeebopConfig = Record<string, any>;
+export interface BeebopRoutes {
+    addRoutes: (app: Application, config: BeebopConfig) => void;
+}
